python/requirements.txt
requests>=2.32.0,<3.0.0
pycryptodome>=3.20.0,<4.0.0
